bigQuery/py/pyproject.toml
[tool.poetry]
name = "davidkhala.gcp.bq"
version = "0.0.0"
description = ""
authors = ["David Liu <david-khala@hotmail.com>"]
readme = "README.md"
packages = [{ include = "davidkhala" }]

[tool.poetry.dependencies]
python = "^3.10"

davidkhala-gcp = "*"
google-cloud-bigquery = [
    { markers = "extra == 'stream'", version = "*", extras = ["bqstorage"] },
    { markers = "extra != 'stream'", version = "*" }
]


# For extras
davidkhala-utils = { version = "*", optional = true }
google-cloud-bigquery-migration = { version = "*", optional = true }

[tool.poetry.group.dev.dependencies]
pytest = "*"
davidkhala-gcp = { version = "*", extras = ["auth"] }

[tool.poetry.extras]
migration = ['google-cloud-bigquery-migration', 'davidkhala-utils']
stream = ['google-cloud-bigquery']
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
